Actualizar id para bigint en versión 5.8 de laravel

+++ src/Migrations/2018_03_22_150317_create_sirgrimorum_paymentpass_payment_passes.php
@@ -13,7 +13,7 @@ class CreateSirgrimorumPaymentpassPaymentPasses extends Migration {
     public function up() {
         Schema::create('payment_passes', function($table) {
             $table->engine = 'InnoDB';
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('referenceCode', 255)->nullable();
             $table->string('type', 3)->nullable();
             $table->string('state', 3)->default('reg');
@@ -26,7 +26,7 @@ class CreateSirgrimorumPaymentpassPaymentPasses extends Migration {
             $table->longtext('response_data')->nullable();
             $table->longtext('confirmation_data')->nullable();
             $table->longtext('creation_data')->nullable();
-            $table->integer('process_id')->unsigned()->nullable();
+            $table->bigInteger('process_id')->unsigned()->nullable();
             $table->timestamps();
         });
     }
